Added condition to detect the dataset in current directory

+++ code.py
@@ -7,12 +7,20 @@ import shutil
 import kagglehub
 
 csv_name = "diabetes_dataset.csv"
-path = kagglehub.dataset_download("mohankrishnathalla/diabetes-health-indicators-dataset")
-print("Path to dataset files:", path)
 
-# move dataset to local directory
-path = shutil.move(path+"/"+csv_name, ".")
-print("New path to dataset files:", path)
+# Download dataset if it doesn't exist
+if (not os.path.exists(csv_name)):
+    path = kagglehub.dataset_download("mohankrishnathalla/diabetes-health-indicators-dataset")
+    print("Path to dataset files:", path)
+
+    # move dataset to local directory
+    try:
+        path = shutil.move(path+"/"+csv_name, ".")
+        print("New path to dataset files:", path)
+    except:
+        print("Error moving dataset files")
+else:
+    print("Dataset already exists in current directory")
 
 # Read the data
 df = pd.read_csv(csv_name)
